feat: improve the implementation by adding the correct merge files
