--- app/src/main/java/ch/hikemate/app/ui/map/HikeDetailScreen.kt
@@ -1,598 +1,0 @@
-package ch.hikemate.app.ui.map
-
-import android.annotation.SuppressLint
-import android.util.Log
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.safeDrawingPadding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.BottomSheetScaffold
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.DatePicker
-import androidx.compose.material3.DatePickerDialog
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.material3.rememberBottomSheetScaffoldState
-import androidx.compose.material3.rememberDatePickerState
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.viewinterop.AndroidView
-import androidx.core.content.ContextCompat
-import androidx.lifecycle.viewmodel.compose.viewModel
-import ch.hikemate.app.R
-import ch.hikemate.app.model.authentication.AuthViewModel
-import ch.hikemate.app.model.facilities.FacilitiesViewModel
-import ch.hikemate.app.model.facilities.Facility
-import ch.hikemate.app.model.profile.HikingLevel
-import ch.hikemate.app.model.profile.ProfileViewModel
-import ch.hikemate.app.model.route.Bounds
-import ch.hikemate.app.model.route.DetailedHikeRoute
-import ch.hikemate.app.model.route.ListOfHikeRoutesViewModel
-import ch.hikemate.app.model.route.saved.SavedHikesViewModel
-import ch.hikemate.app.ui.components.AsyncStateHandler
-import ch.hikemate.app.ui.components.BackButton
-import ch.hikemate.app.ui.components.BigButton
-import ch.hikemate.app.ui.components.ButtonType
-import ch.hikemate.app.ui.components.DetailRow
-import ch.hikemate.app.ui.components.ElevationGraph
-import ch.hikemate.app.ui.components.ElevationGraphStyleProperties
-import ch.hikemate.app.ui.map.HikeDetailScreen.MAP_MAX_ZOOM
-import ch.hikemate.app.ui.map.HikeDetailScreen.TEST_TAG_ADD_DATE_BUTTON
-import ch.hikemate.app.ui.map.HikeDetailScreen.TEST_TAG_BOOKMARK_ICON
-import ch.hikemate.app.ui.map.HikeDetailScreen.TEST_TAG_DATE_PICKER
-import ch.hikemate.app.ui.map.HikeDetailScreen.TEST_TAG_DATE_PICKER_CANCEL_BUTTON
-import ch.hikemate.app.ui.map.HikeDetailScreen.TEST_TAG_DATE_PICKER_CONFIRM_BUTTON
-import ch.hikemate.app.ui.map.HikeDetailScreen.TEST_TAG_ELEVATION_GRAPH
-import ch.hikemate.app.ui.map.HikeDetailScreen.TEST_TAG_HIKE_NAME
-import ch.hikemate.app.ui.map.HikeDetailScreen.TEST_TAG_MAP
-import ch.hikemate.app.ui.map.HikeDetailScreen.TEST_TAG_PLANNED_DATE_TEXT_BOX
-import ch.hikemate.app.ui.map.HikeDetailScreen.TEST_TAG_RUN_HIKE_BUTTON
-import ch.hikemate.app.ui.navigation.NavigationActions
-import ch.hikemate.app.ui.navigation.Route
-import ch.hikemate.app.ui.navigation.Screen
-import ch.hikemate.app.utils.MapUtils
-import ch.hikemate.app.utils.humanReadableFormat
-import com.google.firebase.Timestamp
-import java.util.Date
-import java.util.Locale
-import kotlin.math.max
-import kotlin.math.min
-import kotlin.math.roundToInt
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.combine
-import kotlinx.coroutines.flow.debounce
-import kotlinx.coroutines.flow.launchIn
-import org.osmdroid.events.MapListener
-import org.osmdroid.events.ScrollEvent
-import org.osmdroid.events.ZoomEvent
-import org.osmdroid.util.GeoPoint
-import org.osmdroid.views.CustomZoomButtonsController
-import org.osmdroid.views.MapView
-import org.osmdroid.views.overlay.Marker
-
-object HikeDetailScreen {
-  const val MAP_MAX_ZOOM = 18.0
-  const val MAP_MAX_LONGITUDE = 180.0
-  const val MAP_MIN_LONGITUDE = -180.0
-  const val MAP_BOUNDS_MARGIN: Int = 100
-  const val MARGIN_BOUNDS = 0.001
-
-  const val MIN_ZOOM_FOR_FACILITIES = 15.0
-  const val DEBOUNCE_DURATION = 300L
-
-  const val TEST_TAG_MAP = "HikeDetailScreenMap"
-  const val TEST_TAG_HIKE_NAME = "HikeDetailScreenHikeName"
-  const val TEST_TAG_BOOKMARK_ICON = "HikeDetailScreenBookmarkIcon"
-  const val TEST_TAG_ELEVATION_GRAPH = "HikeDetailScreenElevationGraph"
-  const val TEST_TAG_ADD_DATE_BUTTON = "HikeDetailScreenAddDateButton"
-  const val TEST_TAG_PLANNED_DATE_TEXT_BOX = "HikeDetailScreenPlannedDateTextBox"
-  const val TEST_TAG_DATE_PICKER = "HikeDetailDatePicker"
-  const val TEST_TAG_DATE_PICKER_CANCEL_BUTTON = "HikeDetailDatePickerCancelButton"
-  const val TEST_TAG_DATE_PICKER_CONFIRM_BUTTON = "HikeDetailDatePickerConfirmButton"
-  const val TEST_TAG_RUN_HIKE_BUTTON = "HikeDetailRunHikeButton"
-}
-
-@Composable
-fun HikeDetailScreen(
-    listOfHikeRoutesViewModel: ListOfHikeRoutesViewModel,
-    savedHikesViewModel: SavedHikesViewModel,
-    profileViewModel: ProfileViewModel = viewModel(factory = ProfileViewModel.Factory),
-    authViewModel: AuthViewModel,
-    navigationActions: NavigationActions,
-    facilitiesViewModel: FacilitiesViewModel
-) {
-
-  val context = LocalContext.current
-
-  LaunchedEffect(listOfHikeRoutesViewModel.selectedHikeRoute.collectAsState().value) {
-    if (listOfHikeRoutesViewModel.selectedHikeRoute.value == null) {
-      navigationActions.goBack()
-    }
-  }
-
-  if (listOfHikeRoutesViewModel.selectedHikeRoute.collectAsState().value == null) {
-    Log.e("HikeDetailScreen", "No selected hike route")
-    return
-  }
-
-  val route = listOfHikeRoutesViewModel.selectedHikeRoute.collectAsState().value!!
-  val elevationData = remember { mutableListOf<Double>() }
-
-  val detailedRoute = DetailedHikeRoute.create(route)
-  val routeZoomLevel = MapUtils.calculateBestZoomLevel(route.bounds).toDouble()
-
-  val scope = rememberCoroutineScope()
-  val facilities = remember { mutableStateOf<List<Facility>>(emptyList()) }
-  val facilitiesDisplayed = remember { mutableStateOf(false) }
-
-  // Only do the configuration on the first composition, not on every recomposition
-  LaunchedEffect(Unit) {
-    val bounds = route.bounds
-    val boundsWithMargin =
-        Bounds(
-            (bounds.minLat - HikeDetailScreen.MARGIN_BOUNDS),
-            (bounds.minLon - HikeDetailScreen.MARGIN_BOUNDS),
-            (bounds.maxLat + HikeDetailScreen.MARGIN_BOUNDS),
-            (bounds.maxLon + HikeDetailScreen.MARGIN_BOUNDS))
-    savedHikesViewModel.loadSavedHikes()
-    savedHikesViewModel.updateHikeDetailState(route)
-    // Retrieve ALL the facilities that are inside the boundsWithMargin
-    facilitiesViewModel.getFacilities(boundsWithMargin, { facilities.value = it }, {})
-    listOfHikeRoutesViewModel.getRoutesElevation(
-        route,
-        {
-          elevationData.clear()
-          elevationData.addAll(it)
-        })
-  }
-
-  // Avoid re-creating the MapView on every recomposition
-  val mapView = remember {
-    MapView(context).apply {
-      // Set map's initial state
-      controller.setZoom(routeZoomLevel)
-      controller.setCenter(MapUtils.getGeographicalCenter(route.bounds))
-      // Limit the zoom to avoid the user zooming out or out too much
-      minZoomLevel = routeZoomLevel
-      maxZoomLevel = MAP_MAX_ZOOM
-      // Avoid repeating the map when the user reaches the edge or zooms out
-      // We keep the horizontal repetition enabled to allow the user to scroll the map
-      // horizontally without limits (from Asia to America, for example)
-      isHorizontalMapRepetitionEnabled = true
-      isVerticalMapRepetitionEnabled = false
-      // Disable built-in zoom controls since we have our own
-      zoomController.setVisibility(CustomZoomButtonsController.Visibility.NEVER)
-      // Enable touch-controls such as pinch to zoom
-      setMultiTouchControls(true)
-    }
-  }
-
-  val boundingBoxState = remember { MutableStateFlow(mapView.boundingBox) }
-  val zoomLevelState = remember { MutableStateFlow(mapView.zoomLevelDouble) }
-
-  // This makes it so that only the facilities that are inside the current bounds are displayed
-  LaunchedEffect(mapView) {
-    combine(
-            // Debounce makes it so that this isn't called every time the bounding box changes
-            // but with a slight delay.
-            boundingBoxState.debounce(HikeDetailScreen.DEBOUNCE_DURATION),
-            zoomLevelState) { boundingBox, zoomLevel ->
-              // If the zoomLevel is not high enough it won't display the facilities.
-              // This makes the app run smoother since if this didn't exist there would be too much
-              // facilities in crowded cities like Geneva
-              if (zoomLevel > HikeDetailScreen.MIN_ZOOM_FOR_FACILITIES) {
-
-                MapUtils.displayFacilities(
-                    // Filter all the facilities that are not contained in the boundingBox
-                    facilities.value.filter { facility ->
-                      boundingBox.contains(
-                          GeoPoint(facility.coordinates.lat, facility.coordinates.lon))
-                    },
-                    mapView,
-                    context)
-
-                facilitiesDisplayed.value = true
-              } else if (facilitiesDisplayed.value) {
-                // Clear facilities from the map when zoom level is below the threshold
-                mapView.overlays.removeAll { overlay ->
-                  overlay is Marker && overlay.relatedObject == R.string.facility_marker
-                }
-                facilitiesDisplayed.value = false
-              }
-            }
-        .launchIn(scope)
-  }
-
-  // Update bounding box and zoom level states on map events
-  mapView.addMapListener(
-      object : MapListener {
-        override fun onScroll(event: ScrollEvent?): Boolean {
-          // Whenever the user scrolls this is called to change the boundingBox.
-          event?.let { boundingBoxState.value = mapView.boundingBox }
-          return true
-        }
-
-        override fun onZoom(event: ZoomEvent?): Boolean {
-          // Whenever the user zooms this is called to change the zoomLevel.
-          event?.let { zoomLevelState.value = mapView.zoomLevelDouble }
-          return true
-        }
-      })
-
-  // When the map is ready, it will have computed its bounding box
-  mapView.addOnFirstLayoutListener { _, _, _, _, _ ->
-    // Limit the vertical scrollable area to avoid the user scrolling too far from the hike
-    mapView.setScrollableAreaLimitLatitude(
-        min(MapScreen.MAP_MAX_LATITUDE, mapView.boundingBox.latNorth),
-        max(MapScreen.MAP_MIN_LATITUDE, mapView.boundingBox.latSouth),
-        HikeDetailScreen.MAP_BOUNDS_MARGIN)
-    if (route.bounds.maxLon < HikeDetailScreen.MAP_MAX_LONGITUDE ||
-        route.bounds.minLon > HikeDetailScreen.MAP_MIN_LONGITUDE) {
-      mapView.setScrollableAreaLimitLongitude(
-          max(HikeDetailScreen.MAP_MIN_LONGITUDE, mapView.boundingBox.lonWest),
-          min(HikeDetailScreen.MAP_MAX_LONGITUDE, mapView.boundingBox.lonEast),
-          HikeDetailScreen.MAP_BOUNDS_MARGIN)
-    }
-  }
-
-  // Show the selected hike on the map
-  // OnLineClick does nothing, the line should not be clickable
-  val hikeLineColor = route.getColor()
-  Log.d("HikeDetailScreen", "Drawing hike on map: ${route.bounds}")
-  MapUtils.showHikeOnMap(mapView = mapView, hike = route, color = hikeLineColor, onLineClick = {})
-
-  LaunchedEffect(Unit) {
-    if (authViewModel.currentUser.value == null) {
-      Log.e("HikeDetailScreen", "User is not signed in")
-      return@LaunchedEffect
-    }
-    profileViewModel.getProfileById(authViewModel.currentUser.value!!.uid)
-  }
-
-  var wantToRunHike by remember { mutableStateOf(false) }
-  LaunchedEffect(wantToRunHike) { if (wantToRunHike) navigationActions.navigateTo(Screen.RUN_HIKE) }
-
-  val errorMessageIdState = profileViewModel.errorMessageId.collectAsState()
-  val profileState = profileViewModel.profile.collectAsState()
-
-  AsyncStateHandler(
-      errorMessageIdState = errorMessageIdState,
-      actionContentDescriptionStringId = R.string.go_back,
-      actionOnErrorAction = { navigationActions.navigateTo(Route.MAP) },
-      valueState = profileState,
-  ) { profile ->
-    Box(modifier = Modifier.fillMaxSize().testTag(Screen.HIKE_DETAILS)) {
-      // Map
-      AndroidView(
-          factory = { mapView },
-          modifier =
-              Modifier.fillMaxWidth()
-                  .padding(bottom = 300.dp) // Reserve space for the scaffold at the bottom
-                  .testTag(TEST_TAG_MAP))
-      // Back Button at the top of the screen
-      BackButton(
-          navigationActions = navigationActions,
-          modifier = Modifier.padding(start = 16.dp, end = 16.dp).safeDrawingPadding(),
-          onClick = { listOfHikeRoutesViewModel.clearSelectedRoute() })
-      // Zoom buttons at the bottom right of the screen
-      ZoomMapButton(
-          onZoomIn = { mapView.controller.zoomIn() },
-          onZoomOut = { mapView.controller.zoomOut() },
-          modifier =
-              Modifier.align(Alignment.BottomEnd)
-                  .padding(bottom = MapScreen.BOTTOM_SHEET_SCAFFOLD_MID_HEIGHT + 8.dp))
-
-      // Hike Details bottom sheet
-      HikeDetails(
-          detailedRoute = detailedRoute,
-          savedHikesViewModel = savedHikesViewModel,
-          elevationData = elevationData,
-          userHikingLevel = profile.hikingLevel,
-          onRunThisHike = { wantToRunHike = true },
-      )
-    }
-  }
-}
-
-/**
- * A composable that displays details about a hike in a bottom sheet.
- *
- * @param detailedRoute The route for which information is displayed
- * @param savedHikesViewModel ViewModel managing saved hikes state
- * @param elevationData List of elevation points for the elevation graph
- * @param userHikingLevel The user's hiking experience level
- * @param onRunThisHike Callback triggered when "Run This Hike" button is clicked
- */
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun HikeDetails(
-    detailedRoute: DetailedHikeRoute,
-    savedHikesViewModel: SavedHikesViewModel,
-    elevationData: List<Double>,
-    userHikingLevel: HikingLevel,
-    onRunThisHike: () -> Unit
-) {
-  val hikeDetailState = savedHikesViewModel.hikeDetailState.collectAsState(null).value
-
-  // Handle save/unsave actions
-  val isSaved = hikeDetailState?.isSaved ?: false
-  val toggleSaveState = { savedHikesViewModel.toggleSaveState() }
-
-  // Handle planned date updates
-  val plannedDate = hikeDetailState?.plannedDate
-  val updatePlannedDate = { timestamp: Timestamp? ->
-    savedHikesViewModel.updatePlannedDate(timestamp)
-  }
-
-  val hikeColor = Color(detailedRoute.route.getColor())
-
-  val scaffoldState = rememberBottomSheetScaffoldState()
-
-  val isSuitable = detailedRoute.difficulty.ordinal <= userHikingLevel.ordinal
-
-  BottomSheetScaffold(
-      scaffoldState = scaffoldState,
-      sheetContainerColor = MaterialTheme.colorScheme.surface,
-      sheetContent = {
-        Column(
-            modifier = Modifier.fillMaxSize().padding(16.dp),
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-          Row {
-            Column(
-                modifier = Modifier.padding(16.dp).weight(1f),
-            ) {
-              Text(
-                  text =
-                      detailedRoute.route.name
-                          ?: stringResource(R.string.map_screen_hike_title_default),
-                  style = MaterialTheme.typography.titleLarge,
-                  textAlign = TextAlign.Left,
-                  modifier = Modifier.testTag(TEST_TAG_HIKE_NAME))
-              AppropriatenessMessage(isSuitable)
-            }
-
-            Image(
-                painter = painterResource(hikeDetailState?.bookmark ?: R.drawable.bookmark_no_fill),
-                contentDescription =
-                    if (hikeDetailState?.isSaved == true)
-                        stringResource(R.string.hike_detail_screen_bookmark_hint_on_isSaved_true)
-                    else stringResource(R.string.hike_detail_screen_bookmark_hint_on_isSaved_false),
-                modifier =
-                    Modifier.size(60.dp, 80.dp).testTag(TEST_TAG_BOOKMARK_ICON).clickable {
-                      toggleSaveState()
-                    },
-                contentScale = ContentScale.FillBounds,
-            )
-          }
-
-          ElevationGraph(
-              elevations = elevationData,
-              modifier =
-                  Modifier.fillMaxWidth()
-                      .height(60.dp)
-                      .padding(4.dp)
-                      .testTag(TEST_TAG_ELEVATION_GRAPH),
-              styleProperties =
-                  ElevationGraphStyleProperties(
-                      strokeColor = hikeColor, fillColor = hikeColor.copy(0.1f)))
-
-          val distanceString =
-              String.format(Locale.getDefault(), "%.2f", detailedRoute.totalDistance)
-          val elevationGainString = detailedRoute.elevationGain.roundToInt().toString()
-          val hourString =
-              String.format(Locale.getDefault(), "%02d", (detailedRoute.estimatedTime / 60).toInt())
-          val minuteString =
-              String.format(
-                  Locale.getDefault(), "%02d", (detailedRoute.estimatedTime % 60).roundToInt())
-
-          DetailRow(
-              label = stringResource(R.string.hike_detail_screen_label_distance),
-              value = "${distanceString}km")
-          DetailRow(
-              label = stringResource(R.string.hike_detail_screen_label_elevation_gain),
-              value = "${elevationGainString}m")
-          DetailRow(
-              label = stringResource(R.string.hike_detail_screen_label_estimated_time),
-              value =
-                  if (detailedRoute.estimatedTime / 60 < 1) "${minuteString}min"
-                  else "${hourString}h${minuteString}")
-          DetailRow(
-              label = stringResource(R.string.hike_detail_screen_label_difficulty),
-              value = stringResource(detailedRoute.difficulty.nameResourceId),
-              valueColor =
-                  Color(
-                      ContextCompat.getColor(
-                          LocalContext.current, detailedRoute.difficulty.colorResourceId)),
-          )
-          DateDetailRow(isSaved, plannedDate, updatePlannedDate)
-
-          // "Run This Hike" button
-          BigButton(
-              buttonType = ButtonType.PRIMARY,
-              label = stringResource(R.string.hike_detail_screen_run_this_hike_button_label),
-              onClick = { onRunThisHike() },
-              modifier =
-                  Modifier.padding(top = 16.dp).fillMaxWidth().testTag(TEST_TAG_RUN_HIKE_BUTTON))
-        }
-      },
-      sheetPeekHeight = MapScreen.BOTTOM_SHEET_SCAFFOLD_MID_HEIGHT) {}
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@SuppressLint("StateFlowValueCalledInComposition")
-@Composable
-fun DateDetailRow(
-    isSaved: Boolean,
-    plannedDate: Timestamp?,
-    updatePlannedDate: (Timestamp?) -> Unit
-) {
-  val showingDatePicker = remember { mutableStateOf(false) }
-  val datePickerState = rememberDatePickerState()
-
-  fun showDatePicker() {
-    showingDatePicker.value = true
-  }
-
-  fun dismissDatePicker() {
-    showingDatePicker.value = false
-  }
-
-  if (showingDatePicker.value) {
-    DatePickerDialog(
-        modifier = Modifier.testTag(TEST_TAG_DATE_PICKER),
-        onDismissRequest = { dismissDatePicker() },
-        dismissButton = {
-          Button(
-              modifier = Modifier.testTag(TEST_TAG_DATE_PICKER_CANCEL_BUTTON),
-              onClick = { dismissDatePicker() }) {
-                Text(text = stringResource(R.string.hike_detail_screen_date_picker_cancel_button))
-              }
-        },
-        confirmButton = {
-          Button(
-              modifier = Modifier.testTag(TEST_TAG_DATE_PICKER_CONFIRM_BUTTON),
-              onClick = {
-                if (datePickerState.selectedDateMillis != null) {
-                  updatePlannedDate(Timestamp(Date(datePickerState.selectedDateMillis!!)))
-                }
-
-                dismissDatePicker()
-              }) {
-                Text(text = stringResource(R.string.hike_detail_screen_date_picker_confirm_button))
-              }
-        },
-    ) {
-      DatePicker(state = datePickerState)
-    }
-  }
-
-  if (isSaved) {
-    if (plannedDate == null) {
-      DetailRow(
-          label = stringResource(R.string.hike_detail_screen_label_status),
-          value = stringResource(R.string.hike_detail_screen_value_saved),
-          valueColor = Color(0xFF3B9DE8))
-      Row(
-          modifier = Modifier.fillMaxWidth(),
-          horizontalArrangement = Arrangement.SpaceBetween,
-      ) {
-        Text(
-            text = stringResource(R.string.hike_detail_screen_label_planned_for),
-            style = MaterialTheme.typography.bodyLarge,
-            modifier = Modifier.testTag(DetailRow.TEST_TAG_DETAIL_ROW_TAG))
-
-        Button(
-            modifier = Modifier.width(90.dp).height(25.dp).testTag(TEST_TAG_ADD_DATE_BUTTON),
-            contentPadding = PaddingValues(0.dp),
-            colors =
-                ButtonDefaults.buttonColors(
-                    containerColor = Color(0xFF4285F4), // Blue color to match the image
-                    contentColor = Color.White),
-            onClick = { showDatePicker() },
-        ) {
-          Text(
-              text = stringResource(R.string.hike_detail_screen_add_a_date_button_text),
-              style = MaterialTheme.typography.bodySmall,
-              fontWeight = FontWeight.Bold,
-          )
-        }
-      }
-    } else { // A Date is set
-      DetailRow(
-          label = stringResource(R.string.hike_detail_screen_label_status),
-          value = stringResource(R.string.hike_detail_screen_value_planned),
-          valueColor = Color(0xFF3B9DE8))
-      Row(
-          modifier = Modifier.fillMaxWidth(),
-          horizontalArrangement = Arrangement.SpaceBetween,
-      ) {
-        Text(
-            text = stringResource(R.string.hike_detail_screen_label_planned_for),
-            style = MaterialTheme.typography.bodyLarge,
-            modifier = Modifier.testTag(DetailRow.TEST_TAG_DETAIL_ROW_TAG))
-        Box(
-            modifier =
-                Modifier.border(
-                        BorderStroke(1.dp, MaterialTheme.colorScheme.primary),
-                        shape = RoundedCornerShape(4.dp))
-                    .padding(horizontal = 8.dp, vertical = 4.dp)) {
-              Text(
-                  text = plannedDate.humanReadableFormat(),
-                  // saved Date
-                  style = MaterialTheme.typography.bodySmall,
-                  modifier =
-                      Modifier.clickable { showDatePicker() }
-                          .testTag(TEST_TAG_PLANNED_DATE_TEXT_BOX))
-            }
-      }
-    }
-  } else {
-    DetailRow(
-        label = stringResource(R.string.hike_detail_screen_label_status),
-        value = stringResource(R.string.hike_detail_screen_value_not_saved),
-        valueColor = MaterialTheme.colorScheme.onSurface)
-  }
-}
-
-@Composable
-fun AppropriatenessMessage(isSuitable: Boolean) {
-  // The text, icon and color of the card's message are chosen based on whether the hike is suitable
-  // or not
-  val suitableLabelColor = if (isSuitable) Color(0xFF4CAF50) else Color(0xFFFFC107)
-
-  val suitableLabelText =
-      if (isSuitable) LocalContext.current.getString(R.string.map_screen_suitable_hike_label)
-      else LocalContext.current.getString(R.string.map_screen_challenging_hike_label)
-
-  val suitableLabelIcon =
-      painterResource(if (isSuitable) R.drawable.check_circle else R.drawable.warning)
-
-  Row(verticalAlignment = Alignment.CenterVertically) {
-    Icon(
-        painter = suitableLabelIcon,
-        // The icon is only decorative, the following message is enough for accessibility
-        contentDescription = null,
-        tint = suitableLabelColor,
-        modifier = Modifier.size(16.dp))
-    Spacer(modifier = Modifier.width(4.dp))
-    Text(
-        text = suitableLabelText,
-        style = MaterialTheme.typography.bodySmall,
-        color = suitableLabelColor)
-  }
-}
